Fix error args on bad enum
The Lua error function takes two args: the error string and an integer level. This means the message needs to be concatenated rather than comma-separated. In this case on bad enums we get `bad argument #2 to 'error' (number expected, got string)` when we should really get the message.

Just use concatenation on the string and handle when `name` can be nil.